fastapi/backend/requirements.txt
fastapi>=0.111,<1.0
pandas>=2.2,<3.0
pydantic>=2.12,<3.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
scikit-learn==1.7.2
uvicorn[standard]>=0.30,<1.0